+++ repository/notif_repo.go
@@ -6,9 +6,9 @@ import (
 )
 
 type NotifRepository interface {
-	GetList(userID uint, page, limit int) ([]schema.Notification, int64, error)
-	MarkAsRead(notifID uint, userID uint) error
-	Delete(notifID uint, userID uint) error
+	GetList(userID uint64, page, limit int) ([]schema.Notification, int64, error)
+	MarkAsRead(notifID uint64, userID uint64) error
+	Delete(notifID uint64, userID uint64) error
 	Create(notif *schema.Notification) error
 }
 
@@ -21,7 +21,7 @@ func NewNotifRepository(db *gorm.DB) NotifRepository {
 }
 
 // [71] Lấy danh sách thông báo phân trang
-func (r *notifRepository) GetList(userID uint, page, limit int) ([]schema.Notification, int64, error) {
+func (r *notifRepository) GetList(userID uint64, page, limit int) ([]schema.Notification, int64, error) {
 	var notifs []schema.Notification
 	var total int64
 	offset := (page - 1) * limit
@@ -35,14 +35,14 @@ func (r *notifRepository) GetList(userID uint, page, limit int) ([]schema.Notifi
 }
 
 // [72] Đánh dấu đã đọc
-func (r *notifRepository) MarkAsRead(notifID uint, userID uint) error {
+func (r *notifRepository) MarkAsRead(notifID uint64, userID uint64) error {
 	return r.db.Model(&schema.Notification{}).
 		Where("notif_id = ? AND user_id = ?", notifID, userID).
 		Updates(map[string]interface{}{"is_read": true, "read_at": gorm.Expr("CURRENT_TIMESTAMP")}).Error
 }
 
 // [73] Xóa thông báo
-func (r *notifRepository) Delete(notifID uint, userID uint) error {
+func (r *notifRepository) Delete(notifID uint64, userID uint64) error {
 	return r.db.Where("notif_id = ? AND user_id = ?", notifID, userID).Delete(&schema.Notification{}).Error
 }
